carlisting/cars/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model
from .models import Car

User = get_user_model()


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email')


class CarSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.username')

    class Meta:
        model = Car
        fields = ('id', 'name', 'year', 'price', 'image', 'owner')
