carlisting/requirements.txt
Django==4.0.0
djangorestframework==3.12.4
psycopg2-binary==2.9.3
dj-database-url==0.5.0
django-redis==5.2.0
drf-yasg
